Moves classes to separate files
