+++ Corgibytes.Freshli.Cli/CommandOptions/CacheCommandOptions.cs
@@ -4,12 +4,3 @@ namespace Corgibytes.Freshli.Cli.CommandOptions;
 public class CacheCommandOptions : CommandOptions
 {
 }
-
-public class CachePrepareCommandOptions : CommandOptions
-{
-}
-
-public class CacheDestroyCommandOptions : CommandOptions
-{
-    public bool Force { get; set; }
-}
