+++ Corgibytes.Freshli.Cli/CommandOptions/CacheDestroyCommandOptions.cs
@@ -1,0 +1,6 @@
+﻿namespace Corgibytes.Freshli.Cli.CommandOptions;
+
+public class CacheDestroyCommandOptions : CommandOptions
+{
+    public bool Force { get; set; }
+}
