+++ Corgibytes.Freshli.Cli/CommandOptions/CachePrepareCommandOptions.cs
@@ -1,0 +1,5 @@
+﻿namespace Corgibytes.Freshli.Cli.CommandOptions;
+
+public class CachePrepareCommandOptions : CommandOptions
+{
+}
