+++ Corgibytes.Freshli.Cli/DataModel/CacheContext.cs
@@ -1,7 +1,5 @@
 using System.IO;
-using Corgibytes.Freshli.Cli.Functionality;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Design;
 
 namespace Corgibytes.Freshli.Cli.DataModel;
 
@@ -26,18 +24,4 @@ public class CacheContext : DbContext
 
     protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
         => optionsBuilder.UseSqlite($"Data Source={DbPath}");
-}
-
-// ReSharper disable once UnusedType.Global
-public class CacheContextFactory : IDesignTimeDbContextFactory<CacheContext>
-{
-    public CacheContext CreateDbContext(string[] args)
-    {
-        var configuration = new Configuration(new Environment());
-        var optionsBuilder = new DbContextOptionsBuilder<CacheContext>();
-        var dbPath = Path.Join(configuration.CacheDir, CacheContext.CacheDbName);
-        optionsBuilder.UseSqlite($"Data Source={dbPath}");
-
-        return new CacheContext(configuration.CacheDir);
-    }
 }
